Fixed issue with Instalfred not working with browsing running concurrently
Instalfred now plays nicely with multiple browsers running at once. It
now considers what the frontmost application is, not just what
applications are running (this caused an issue where it would always
grab the URL from Safari first if it was running).

Also added an error message when Instalfred is called on a non-browser
window.